requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# OCR and image processing
pytesseract==0.3.10
opencv-python-headless==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# Data validation and parsing
pydantic==2.4.2

# Logging and utilities
python-json-logger==2.0.7

# Development and testing (optional)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Production server (alternative to uvicorn)
gunicorn==21.2.0